Split some files and include fetch mock
