+++ package.json
@@ -1,82 +1,86 @@
 {
-  "name": "tubular-common",
-  "version": "2.0.3",
-  "description": "Tubular Common Models and Data Transformer",
-  "main": "dist/index.js",
-  "scripts": {
-    "test": "jest --runInBand",
-    "lint": "eslint src/**/*.{js,ts}",
-    "lint:fix": "eslint src/**/*.{js,ts} --fix",
-    "prettier": "prettier --check src/**/*.ts*",
-    "prettier:fix": "prettier --write src/**/*.ts*",
-    "codestyle:fix": "npm run lint:fix && npm run prettier:fix",
-    "build": "tsc",
-    "prepublish": "npm run build"
-  },
-  "jest": {
-    "automock": false,
-    "coverageReporters": [
-      "lcov",
-      "text"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "js"
-    ],
-    "transform": {
-      "^.+\\.(ts)$": "ts-jest"
+    "name": "tubular-common",
+    "version": "2.0.3",
+    "description": "Tubular Common Models and Data Transformer",
+    "main": "dist/index.js",
+    "scripts": {
+        "test": "jest --runInBand",
+        "lint": "eslint src/**/*.{js,ts}",
+        "lint:fix": "eslint src/**/*.{js,ts} --fix",
+        "prettier": "prettier --check src/**/*.ts*",
+        "prettier:fix": "prettier --write src/**/*.ts*",
+        "codestyle:fix": "npm run lint:fix && npm run prettier:fix",
+        "build": "tsc",
+        "prepublish": "npm run build"
     },
-    "testMatch": [
-      "<rootDir>/test/**/*.spec.ts"
-    ],
-    "globals": {
-      "ts-jest": {
-        "diagnostics": false,
-        "tsConfig": "tsconfig.json"
-      }
+    "jest": {
+        "automock": false,
+        "setupFiles": [
+            "./setupJest.ts"
+        ],
+        "coverageReporters": [
+            "lcov",
+            "text"
+        ],
+        "moduleFileExtensions": [
+            "ts",
+            "js"
+        ],
+        "transform": {
+            "^.+\\.(ts)$": "ts-jest"
+        },
+        "testMatch": [
+            "<rootDir>/test/**/*.spec.ts"
+        ],
+        "globals": {
+            "ts-jest": {
+                "diagnostics": false,
+                "tsConfig": "tsconfig.json"
+            }
+        },
+        "collectCoverage": true,
+        "coverageThreshold": {
+            "global": {
+                "branches": 45,
+                "functions": 35,
+                "lines": 65,
+                "statements": -160
+            }
+        }
     },
-    "collectCoverage": true,
-    "coverageThreshold": {
-      "global": {
-        "branches": 45,
-        "functions": 35,
-        "lines": 65,
-        "statements": -160
-      }
+    "files": [
+        "dist/**/*",
+        "src/**/*"
+    ],
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/unosquare/tubular-common.git"
+    },
+    "keywords": [
+        "tubular",
+        "data",
+        "grid"
+    ],
+    "author": "Unosquare",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/unosquare/tubular-common/issues"
+    },
+    "homepage": "https://github.com/unosquare/tubular-common#readme",
+    "devDependencies": {
+        "@types/jest": "^25.2.3",
+        "@typescript-eslint/eslint-plugin": "^2.34.0",
+        "@typescript-eslint/parser": "^2.34.0",
+        "eslint": "^7.0.0",
+        "eslint-config-prettier": "^6.11.0",
+        "eslint-plugin-prettier": "^3.1.3",
+        "jest": "^26.0.1",
+        "jest-fetch-mock": "^3.0.3",
+        "prettier": "^2.0.5",
+        "ts-jest": "^26.0.0",
+        "typescript": "^3.9.2"
+    },
+    "dependencies": {
+        "uno-js": "^0.7.22"
     }
-  },
-  "files": [
-    "dist/**/*",
-    "src/**/*"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/unosquare/tubular-common.git"
-  },
-  "keywords": [
-    "tubular",
-    "data",
-    "grid"
-  ],
-  "author": "Unosquare",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/unosquare/tubular-common/issues"
-  },
-  "homepage": "https://github.com/unosquare/tubular-common#readme",
-  "devDependencies": {
-    "@types/jest": "^25.2.3",
-    "@typescript-eslint/eslint-plugin": "^2.34.0",
-    "@typescript-eslint/parser": "^2.34.0",
-    "eslint": "^7.0.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.3",
-    "jest": "^26.0.1",
-    "prettier": "^2.0.5",
-    "ts-jest": "^26.0.0",
-    "typescript": "^3.9.2"
-  },
-  "dependencies": {
-    "uno-js": "^0.7.22"
-  }
 }
