+++ setupJest.ts
@@ -1,0 +1,3 @@
+import fetch from 'jest-fetch-mock';
+
+fetch.enableMocks();
