+++ src/Models/CompareOperator.ts
@@ -1,0 +1,6 @@
+import { CompareOperators } from './Column';
+
+export interface CompareOperator {
+    value: CompareOperators;
+    title: string;
+}
