+++ src/Models/index.ts
@@ -1,4 +1,6 @@
 export * from './Column';
 export * from './ColumnModel';
 export * from './GridRequest';
+export * from './columnUtils';
+export * from './CompareOperator';
 export * from './GridResponse';
